976102201881426068
ManifestID: 976102201881426068
Date: 2026-06-28T22:33:10+08:00
Version: AT-242372

+++ Assembly-CSharp/GraphAxis.cs
@@ -19,5 +19,5 @@ public struct GraphAxis
 
 	private LocText name_label;
 
-	public int guide_frequency;
+	public float guide_frequency;
 }
